Add files via upload
My attempt at this, definitely can be highly optimized but it does a lot of stuff, I don't know any way to select the specific programmer you want, I tried using an array that holds the arrays of all 5 programmers but it didn't work. So I switched to a switch-case code that doesn't work perfectly. I hope to get feedback on how it could improve, thanks so much!!!!